Use CPS for all the control methods.
This enables deeply nested calls to sleep(ms) to somehow
suspend the current control stack, which can be resumed
later. All of this without using a separate thread, so
not giving up hope of porting the whole thing to Scala.js.

Thanks to the continuations plugin of Scala, it is possible
to design the lower levels of the API in a way that users,
including developers of custom squares, need not know a
thing about CPS!