Abstracted away JavaFX from the core and mazes modules.
The packages com.funlabyrinthe.core.{graphics,input} now
define a platform-agnostic, abstract API for drawing
and receiving input.

The package com.funlabyrinthe.graphics.jfx implements
that interface on top of JavaFX, but core and mazes need
not know about that.

This will allow for alternative implementations of the
graphics system, such as on top of HTML5.