New module system.
Modules are now `object`s instead of `class`es. They create their
components in a method `createComponents()(using Universe)`, and
then re-offer access to them as separate accessor methods.

This causes more duplication and left safety for the definition of
a module. However, with an `export TheModule.*`, it allows to very
conveniently *use* all the definitions of a module.

We intend to recover the dryness and safety of module definitions
later with a compiler plugin. The idea is that we will only define
the accessor `def`s, with an rhs that creates the component. The
compiler plugin will then move the rhs to the `createComponents()`
method, and replace it with the unsafe accessor.