Make the full ID of components hierarchical.
Previously, the `id` of a component was a globally unique identifier
in the entire universe. This was problematic for at least two reasons:

* Multiple modules could never use the same ID. For example, a module
  for free-walking maps with grass could not call its grass
  component `grass`, since it would collide the one blessed for Mazes.
* Components that create derived components (like `Vehicle`s and their
  `VehiclePlugin`s) had to craft elaborate IDs for their derived
  components.

Now, IDs are hierarchical. At the top, they are namespaced by the
module. For derived components, which we call sub-components, IDs
are namespaced to their owner component.

Attributes also receive IDs that are namespaced to their enclosing
module. They don't have more levels of nesting, however.