Remove NamedComponent and VisualComponent.
`NamedComponent.name` was never read, so it was useless.

`VisualComponent` was too specific to `Mazes`, whose square size
happens to match icon sizes. Yet at the same time its `drawTo`
methods had the wrong signature for `Mazes`. So we remove it and
directly introduce drawing in `SquareComponent` and `PosComponent`
instead.