Use JS Promise Integration of Wasm to implement controls: minimal changes.
Previously, we used a monadic structure `Control[+A]` to deal with
semi-synchronous waits: `sleep`s and `waitForKeyEvent`. We used
dotty-cps-async to CPS-transform every method on the path to a
potential `sleep` or `waitForKeyEvent`.

While dotty-cps-async works well when it works, it dramatically
obscures compile errors.

Instead, we now use the magic power of the JS Promise Integration
(JSPI) proposal for WebAssembly. This proposal allows Wasm code to
synchronously wait for completion of a JavaScript promise, as long
as the entry point into Wasm returns a `Promise` itself.

We can abuse this system to have truly synchronous calls to `sleep`
and `waitForKeyEvent`: they generate a new JS Promise that will be
completed when the delay is over or when a key is pressed, then
wait for that promise at the boundary between Wasm and JS.

This completely removes the need for dotty-cps-async, and for our
`Control` abstraction.

In this commit, we only perform the minimal changes that switch to
the new system. In the next commit, we will remove `Control` and
all leftover references to dotty-cps-async.