Add support to copy and delete additional components.
This includes a significant redesign of additional components and
component creators. Previously, component creators owned the
components they created. They could also fully tweak how they were
instantiated.

Now, they can only specify the *class* of the components they
create. The actual instantiation is done using reflection, based on
the constructor taking a `ComponentInit`. Moreover, the created
components belong to a special module `AdditionalComponents`. That
module recreates additional components based on their serialzied
class name.

With all that, we have the infrastructure to *copy* components. We
can indeed create them using the same reflection-based mechanism.
They also belong to `AdditionalComponents`, and are serialized in
the same way.

In order to *delete* component, we enhance the `Pickleable` and
`InPlacePickleable` typeclasses. They now have an additional method
to remove references to a given component from their values. When
that is not possible, an error is displayed, and the component
cannot be destroyed.