Redesign the `Reflectable` system.
We get rid of the `Reflector` typeclass. Instead, we directly
enumerate `InspectedData` in `Reflectable.reflectProperties`.
That method should be overridden in all subclasses, and call
`super` + `Reflectable.autoReflectProperties`. The latter method
is a macro that replaces the derivation of `Reflector`.

We add a compiler plugin that automatically inserts that standard
override of `reflectProperties` in all subclasses and subtraits of
`Reflectable`. If there is already a user-defined one, it leaves
it alone. This way, typical component classes never have to worry
about the reflection boilerplate.

Since we call `super`, the logic in `Reflector` only processes
new *declarations*, instead of all *members*. This makes the
output resistent to binary-compatible additions of properties in
superclasses. It will also allow us to pickle private properties
in the future.