Make Abilities pickleable and serializable.
We achieve this with a registration mechanism. It is a bit of a hybrid
between what we did for Attributes and for sum types. Whereas
Attributes are sort of a dynamic struct type, an Ability is a dynamic
sum type.

With that, we complete the capabilities of "simple squares" to be on
par with FunLabyrinthe 5: simple items, simple fields and simple
obstacles, which can communicate through customizable Abilities.